requirements.txt
python-telegram-bot==20.3
sqlalchemy==2.0.35
plotly==5.24.1
kaleido==0.2.1
